refactor: `WaitingValidator` 클래스 삭제
