+++ src/main/java/roomescape/reservation/application/UserWaitingService.java
@@ -1,5 +1,6 @@
 package roomescape.reservation.application;
 
+import java.time.LocalDate;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import roomescape.global.exception.ResourceNotFoundException;
@@ -10,23 +11,24 @@ import roomescape.reservation.application.dto.response.WaitingServiceResponse;
 import roomescape.reservation.model.entity.ReservationTheme;
 import roomescape.reservation.model.entity.ReservationTime;
 import roomescape.reservation.model.entity.Waiting;
+import roomescape.reservation.model.exception.ReservationException.InvalidReservationTimeException;
+import roomescape.reservation.model.repository.ReservationRepository;
 import roomescape.reservation.model.repository.ReservationThemeRepository;
 import roomescape.reservation.model.repository.ReservationTimeRepository;
 import roomescape.reservation.model.repository.WaitingRepository;
-import roomescape.reservation.model.service.WaitingValidator;
 
 @Service
 @RequiredArgsConstructor
 public class UserWaitingService {
 
     private final WaitingRepository waitingRepository;
+    private final ReservationRepository reservationRepository;
     private final ReservationThemeRepository reservationThemeRepository;
     private final ReservationTimeRepository reservationTimeRepository;
     private final MemberRepository memberRepository;
-    private final WaitingValidator waitingValidator;
 
     public WaitingServiceResponse create(CreateWaitingServiceRequest request) {
-        waitingValidator.validateNoDuplication(request.date(), request.timeId(), request.themeId(),
+        validateNoDuplication(request.date(), request.timeId(), request.themeId(),
             request.memberId());
 
         ReservationTheme theme = reservationThemeRepository.getById(request.themeId());
@@ -47,5 +49,14 @@ public class UserWaitingService {
         Waiting waiting = waitingRepository.findByIdAndMemberId(id, memberId)
             .orElseThrow(() -> new ResourceNotFoundException("id에 해당하는 예약 대기가 존재하지 않습니다."));
         waitingRepository.delete(waiting);
+    }
+
+    public void validateNoDuplication(LocalDate date, Long timeId, Long themeId, Long memberId) {
+        if (waitingRepository.existsByDateAndTimeIdAndThemeIdAndMemberId(date, timeId, themeId,
+            memberId) ||
+            reservationRepository.existsByDateAndTimeIdAndThemeIdAndMemberId(date, timeId, themeId,
+                memberId)) {
+            throw new InvalidReservationTimeException("이미 대기중인 예약입니다.");
+        }
     }
 }
